Cluster transformation function
This PR fixes following bug:
If RoleRequest was created and only after Cluster labels were changed
for Cluster to be a match, RoleRequest was not reconciled.

In order to fix bug, keep in memory the ClusterSelector for each
existing RoleRequest.
When cluster changes, requeue:
1. all RoleRequests that used to match such cluster (this was already
working);
2. Walk all existing RoleRequests, requeue each RoleRequest whose
ClusterSelector is now a match for cluster (this is now fixed using new
information kept in memory)